Validate first-party DotSlash artifacts continuously and after release
Summary:
Adds a workflow to ensure the following:

* Continuously: `main` contains valid DotSlash files that reference valid artifacts.
* When a release is published: the release contains valid DotSlash files that reference valid artifacts.

This gives us confidence that branch cuts start in a valid state and that releases go out in a valid state. This is particularly important since there is a span of time, between the release commit and the publishing of the GH release, when the job would fail (as the release assets are not hosted at their final URLs yet).

This is a fairly trivial script built on top of unit-tested utilities: `processDotSlashFileInPlace`, `getWithCurl`, `validateDotSlashArtifactData`.