Add missing AccessibilityProps::role to getDebugProps
Summary:
## Changelog:
[Internal] - 

This was missing, leading to not being able to get the corresponding value back from native for some of the Fantom tests.

I also modified the corresponding AccessibilityProps related Fantom tests, so that we are now testing the right things for all relevant component types (see the test plan).

Differential Revision: D81786901