Add Function to deintegrate local SwiftPM packages from Xcode project
Summary:
## Context

Apple does not have a tool from CLI to add a package to
a project, therefore we need to manually modify the Xcodeproject to integrate some Swift packages into it.

## Changed

We are adding a function that deintegrate local Swift Packages. This helps to make sure we can have an idempotent script and that we don't add the same package twice.

## Changelog:
[Internal] -

Differential Revision: D81778458