Add helper functions to add local packages to xcodeproject
Summary:
## Context

Apple does not have a tool from CLI to add a package to
a project, therefore we need to manually modify the Xcodeproject to integrate some Swift packages into it.

## Changed

This change does two things:
- moves the function that generate the object id to a different file
- adds a function to add the local Swift packages to the project

Moving the generateObjectId function to a separate file is a required step to mock it in the tests.

## Changelog:
[Internal] -

Differential Revision: D81934363