Add function to fix the REACT_NATIVE_PATH variable
Summary:
## Context

When configuring an app to build with SwiftPM from source, there is a sequence of operations we need to run in order to prepare the project correctly.

## Changed

Add a function that fixes the REACT_NATIVE_PATH build settings. Cocoapods sets this value to something that starts with `PODS_ROOT`. However, without using Cocoapods, we can't really rely on that variable being set.

## Changelog:
[Internal] -

Differential Revision: D81778462