Add Functions to coordinate the update of the xcode project.
Summary:
## Context

Apple does not have a tool from CLI to add a package to
a project, therefore we need to manually modify the Xcodeproject to integrate some Swift packages into it.

## Changed

This is the last function required to manipulate the Xcodeproject file. It is an high-level function that orchestrate all the changes that are needed in the xcodeproj file

## Changelog:
[Internal] -

Differential Revision: D82020721