Add Function to add missing sections in case the original project don't have them (#53654)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/53654

Pull Request resolved: https://github.com/facebook/react-native/pull/53648

## Context

Apple does not have a tool from CLI to add a package to
a project, therefore we need to manually modify the Xcodeproject to integrate some Swift packages into it.

## Changed

This change adds a function to add missing section to the project. This function is used by a coordination function in a future diff on the stack.

## Changelog:
[Internal] -

Differential Revision: D81778443