Add script to add the RN Swift packages to xcodeproject (#53668)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/53668

## Context

Apple does not have a tool from CLI to add a package to
a project, therefore we need to manually modify the Xcodeproject to integrate some Swift packages into it.

## Changed

This change adds an higeher level function that can be used to integrate local swift packages into Xcodeproject files.

This will also be helpful for autolinking packages.

## Changelog:
[Internal] -

Differential Revision: D81778453