Add Function to update the package reference section (#53659)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/53659

## Context

Apple does not have a tool from CLI to add a package to
a project, therefore we need to manually modify the Xcodeproject to integrate some Swift packages into it.

## Changed

This change adds a function to update the package reference section so that Xcode can properly load the packages in the project.

## Changelog:
[Internal] -

Differential Revision: D81778463