Add function to generate random Xcode ids (#53643)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/53643

## Context

Apple does not have a tool from CLI to add a package to
a project, therefore we need to manually modify the Xcodeproject to integrate some Swift packages into it.

## Changed

All the Xcode project elements have an id of 24 hex characters. This change adds a function to generate the id randomly.

## Changelog:
[Internal] -

Reviewed By: cortinico

Differential Revision: D81778473