Add helper functions to serialize specific objects of the project (#53647)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/53647

## Context

Apple does not have a tool from CLI to add a package to
a project, therefore we need to manually modify the Xcodeproject to integrate some Swift packages into it.

## Changed

This change adds some helper function that can be used to serialize some simple Xcodeproject objects back to string.
This will be used in future diffs to rewrite the Xcodeproj file.

## Changelog:
[Internal] -

Reviewed By: cortinico

Differential Revision: D81778445