Make RNester build React Native from source with SwiftPM (#53699)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/53699

This change configures RNTester to build from source with SwiftPM.

it introduces a `swiftpm.config.js`, which is basically a replacement for the Podfile file, where users can define manually additional Swift packages that they want to integrate with their apps.

This is the same mechanism of the Podfile. Swift PM does not have a cli to do so. This file will be useful also for autolinking with swiftpm in the future, as autolinking can generate the `swiftpm.config.js` file or we can define an hook for it.

The change also adds a shortcut to run the prepare script.

## Changelog:
[Internal] -

Reviewed By: cortinico

Differential Revision: D81778447