Add Function to convert pbx file to JSON (#53644)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/53644

## Context

Apple does not have a tool from CLI to add a package to
a project, therefore we need to manually modify the Xcodeproject to integrate some Swift packages into it.

## Changed

Apple provides a helpful CLI program called `plutils` that can take a xcodeproj and convert it to json. We use that to parse the xcodeproject.

## Changelog:
[Internal] -

Differential Revision: D81778448