Fix issue with flexShrink
Summary:
In cases where the remaining space for flex shrink is negative, we can get into a state where the `totalFlexShrinkScaledFactors` get's infintesimally small, but likely should just be floored to zero.

This manifests in issues where flexShrink nodes are suddenly given way too much space, and generally only manifests in an issue if these flexShrink nodes have self-measuring children (such as text).

Checking if the value is approximately zero appears to fix bugs for flexShrink.

Differential Revision: D82313674