Inline lazy fields in core data types (#53959)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/53959

These data-types are super common in React Native, and the anonymous inner classes add significantly to the GC workload we see. For ReadableMap this means we can save 6 additional object allocations for each LazySynchronizedImpl and the lambda that's passed in to it. On synthetic benchmarks this seems to improve core operations around ~10%.

Changelog: [Internal]

Reviewed By: lenaic

Differential Revision: D83245864

fbshipit-source-id: 69b314a1f52a32375b5e1f591555f50bfbc31843