[ios][precompiled] add support for symbolication
This commit adds support for symbolication of the XCFrameworks on request.

The reason doing this is that the symbol files are big and only needed if you need to debug React Native itself - f.ex. if you are a framework developer like Expo.

Symbolication can be performed by setting the `RCT_SYMBOLICATE_PREBUILT_FRAMEWORKS` environment variable to the value ''1'. This will cause the `ReactNativeCoreUtils` helper class to download symbol files and symbolicate the XFrameworks by doing the following:

- After downloading the requested React.XCFramework the symbols will also be downloaded and places in the artifacts folder.
- The XCFrameworks will be expanded and the folders in the symbol archive will be extracted into the XCFramework before it is zipped up again.

Some additional parts of this refactoring:
- Aligned logging in both rndependencies.rb and rncore.rb
- Removed uneccessary logging
- RNCore/Deps: Changed so that the source for the package is the locally downloaded file so we don't download twice!